add resource command dispatcher
move resource script related commands into
their own command dispatcher
